Extracted cli to its own module
