+++ src/utils/enums.rs
@@ -1,62 +1,62 @@
-use clap::{Subcommand, ValueEnum};
+// use clap::{Subcommand, ValueEnum};
 
-/// Logging format options
-#[derive(Debug, Clone, Copy, ValueEnum)]
-pub enum LogFormat {
-    /// Plain text format, better for local development
-    Plain,
-    /// JSON format, better for production and machine parsing
-    Json,
-}
+// /// Logging format options
+// #[derive(Debug, Clone, Copy, ValueEnum)]
+// pub enum LogFormat {
+//     /// Plain text format, better for local development
+//     Plain,
+//     /// JSON format, better for production and machine parsing
+//     Json,
+// }
 
-/// Output format options for displaying data
-#[derive(Debug, Clone, Copy, ValueEnum, PartialEq)]
-pub enum OutputFormat {
-    /// Standard output format with essential columns
-    Normal,
-    /// Extended output format with additional columns
-    Wide,
-}
+// /// Output format options for displaying data
+// #[derive(Debug, Clone, Copy, ValueEnum, PartialEq)]
+// pub enum OutputFormat {
+//     /// Standard output format with essential columns
+//     Normal,
+//     /// Extended output format with additional columns
+//     Wide,
+// }
 
-/// CLI command structure
-#[derive(Subcommand, Debug)]
-pub enum Commands {
-    /// Get information about Kubernetes resources
-    Get {
-        #[command(subcommand)]
-        resource: GetImages,
-    },
-}
+// /// CLI command structure
+// #[derive(Subcommand, Debug)]
+// pub enum Commands {
+//     /// Get information about Kubernetes resources
+//     Get {
+//         #[command(subcommand)]
+//         resource: GetImages,
+//     },
+// }
 
-/// Resource types that can be queried
-#[derive(Subcommand, Debug)]
-pub enum GetImages {
-    /// List pod images and their registries
-    Images {
-        /// Kubernetes namespace to query (defaults to "default", ignored when --node is specified)
-        #[arg(
-            short,
-            long,
-            default_value = "default",
-            conflicts_with = "all_namespaces"
-        )]
-        namespace: String,
+// /// Resource types that can be queried
+// #[derive(Subcommand, Debug)]
+// pub enum GetImages {
+//     /// List pod images and their registries
+//     Images {
+//         /// Kubernetes namespace to query (defaults to "default", ignored when --node is specified)
+//         #[arg(
+//             short,
+//             long,
+//             default_value = "default",
+//             conflicts_with = "all_namespaces"
+//         )]
+//         namespace: String,
 
-        #[arg(short = 'N', long = "node", conflicts_with = "all_namespaces")]
-        node: Option<String>,
+//         #[arg(short = 'N', long = "node", conflicts_with = "all_namespaces")]
+//         node: Option<String>,
 
-        #[arg(short, long)]
-        pod: Option<String>,
+//         #[arg(short, long)]
+//         pod: Option<String>,
 
-        /// Filter pods by container image registry
-        #[arg(short = 'R', long = "registry")]
-        registry: Option<String>,
+//         /// Filter pods by container image registry
+//         #[arg(short = 'R', long = "registry")]
+//         registry: Option<String>,
 
-        #[arg(short = 'A', long = "all-namespaces", conflicts_with = "namespace")]
-        all_namespaces: bool,
+//         #[arg(short = 'A', long = "all-namespaces", conflicts_with = "namespace")]
+//         all_namespaces: bool,
 
-        /// Output format (default: normal, wide: shows additional columns)
-        #[arg(short = 'o', long = "output", default_value = "normal")]
-        output: OutputFormat,
-    },
-}
+//         /// Output format (default: normal, wide: shows additional columns)
+//         #[arg(short = 'o', long = "output", default_value = "normal")]
+//         output: OutputFormat,
+//     },
+// }
